fix(gatsby-plugin-clerk): Refactor authenticateRequest to use the new backend pkg
